feat:auth page functional
