+++ src/Futura_frontend/src/components/Authentication.tsx
@@ -1,103 +1,181 @@
-// import { Metadata } from "next";
-// import Image from "next/image";
-//import the image from the public folder
+import { useState, useEffect } from "react";
+import PlugConnect from "@psychedelic/plug-connect";
 import { Link } from "react-router-dom";
-
-// import { cn } from "@/lib/utils";
-// import { buttonVariants } from "@/registry/new-york/ui/button";
-// import { UserAuthForm } from "@/app/(app)/examples/authentication/components/user-auth-form";
-
-// export const metadata: Metadata = {
-//   title: "Authentication",
-//   description: "Authentication forms built using the components.",
-// };
+import { Button } from "@/components/ui/button";
 
 export default function AuthenticationPage() {
+  const [view, setView] = useState("auth");
+  const [isConnected, setIsConnected] = useState(false);
+  const [principal, setPrincipal] = useState(null);
+
+  const privacyPolicyContent = `
+    You know, it's just a hackathon.
+  `;
+
+  const termsOfServiceContent = `
+    We don't have any of those.
+  `;
+
+  const verifyConnectionAndAgent = async () => {
+    const whitelist = [process.env.VITE_CANISTER_ID_FUTURA_BACKEND];
+    const host = "https://mainnet.dfinity.network"; // Adjust the host if necessary
+
+    const connected = await window.ic.plug.isConnected();
+
+    if (!connected) {
+      await window.ic.plug.requestConnect({ whitelist, host });
+    }
+
+    if (connected && !window.ic.plug.agent) {
+      await window.ic.plug.createAgent({ whitelist, host });
+    }
+
+    const principal = window.ic.plug.agent
+      ? await window.ic.plug.agent.getPrincipal()
+      : null;
+    if (principal) {
+      console.log("User principal fetched successfully:", principal.toString());
+      setPrincipal(principal.toString());
+      setIsConnected(true);
+    }
+  };
+
+  useEffect(() => {
+    verifyConnectionAndAgent();
+  }, []);
+
+  function handleDisconnect() {
+    setPrincipal(null);
+    setIsConnected(false);
+    window.ic.plug.agent = null;
+    console.log("Disconnected from Plug");
+  }
+
+  const shortenPrincipal = (principal) => {
+    if (!principal) return "";
+    const start = principal.slice(0, 15);
+    const end = principal.slice(-3);
+    return `${start}...${end}`;
+  };
+
   return (
     <>
-      <div className="md:hidden">
-        <img
-          src="/image.jpeg" // Referencing the image in the public folder
-          width={1280}
-          height={843}
-          alt="Authentication"
-          className="block dark:hidden"
-        />
-        <img
-          src="/image.jpeg" // Using the same image for dark mode as an example
-          width={1280}
-          height={843}
-          alt="Authentication"
-          className="hidden dark:block"
-        />
-      </div>
-      <div className="container relative hidden h-[800px] flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
-        {/* <Link
-          href="/examples/authentication"
-          className={cn(
-            buttonVariants({ variant: "ghost" }),
-            "absolute right-4 top-4 md:right-8 md:top-8"
-          )}
-        >
-          Login
-        </Link> */}
-        <div className="relative hidden h-full flex-col bg-muted p-10 text-white dark:border-r lg:flex">
-          <div className="absolute inset-0 bg-zinc-900" />
-          <div className="relative z-20 flex items-center text-lg font-medium">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="mr-2 h-6 w-6"
-            >
-              <path d="M15 6v12a3 3 0 1 0 3-3H6a3 3 0 1 0 3 3V6a3 3 0 1 0-3 3h12a3 3 0 1 0-3-3" />
-            </svg>
-            Acme Inc
+      {view === "auth" && (
+        <div>
+          <div className="md:hidden">
+            <img
+              src="/image.jpg"
+              width={1280}
+              height={843}
+              alt="Authentication"
+              className="block dark:hidden"
+            />
+            <img
+              src="/image.jpg"
+              width={1280}
+              height={843}
+              alt="Authentication"
+              className="hidden dark:block"
+            />
           </div>
-          <div className="relative z-20 mt-auto">
-            <blockquote className="space-y-2">
-              <p className="text-lg">
-                &ldquo;happiness only real when shared.&rdquo;
-              </p>
-              <footer className="text-sm">Chris - Into The Wild</footer>
-            </blockquote>
-          </div>
-        </div>
-        <div className="lg:p-8">
-          <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
-            <div className="flex flex-col space-y-2 text-center">
-              <h1 className="text-2xl font-semibold tracking-tight">
-                Log-in with Plug
-              </h1>
-              <p className="text-sm text-muted-foreground">
-                Authenticate with your account to access the platform.
-              </p>
+          <div className="container relative hidden h-[800px] flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
+            <div className="relative hidden h-full flex-col bg-muted p-10 text-white dark:border-r lg:flex">
+              <div className="absolute inset-0 bg-zinc-900" />
+              <div className="relative z-20 flex items-center text-lg font-medium">
+                Futura Inc
+              </div>
+              <div className="relative z-20 mt-auto">
+                <blockquote className="space-y-2">
+                  <p className="text-lg">
+                    &ldquo;happiness only real when shared.&rdquo;
+                  </p>
+                  <footer className="text-sm">Chris - Into The Wild</footer>
+                </blockquote>
+              </div>
             </div>
-            {/* <UserAuthForm /> */}
-            <p className="px-8 text-center text-sm text-muted-foreground">
-              By clicking continue, you agree to our{" "}
-              <Link
-                to="/terms"
-                className="underline underline-offset-4 hover:text-primary"
-              >
-                Terms of Service
-              </Link>{" "}
-              and{" "}
-              <Link
-                to="/privacy"
-                className="underline underline-offset-4 hover:text-primary"
-              >
-                Privacy Policy
-              </Link>
-              .
-            </p>
+            <div className="lg:p-8">
+              <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
+                <div className="flex flex-col space-y-2 text-center">
+                  <h1 className="text-2xl font-semibold tracking-tight">
+                    Log-in with Plug
+                  </h1>
+                  <p className="text-sm text-muted-foreground">
+                    Authenticate with your account to access the platform.
+                  </p>
+                </div>
+                {isConnected ? (
+                  <div className="flex flex-col justify-center items-center space-y-4 mt-5">
+                    <div className="flex items-center p-1 bg-green-500 text-white rounded-full transition-colors duration-300 cursor-pointer">
+                      <img
+                        src="/Pluglogo.svg"
+                        alt="Plug logo"
+                        className="w-8 h-8 rounded-full mr-3"
+                      />
+                      <span className="">{shortenPrincipal(principal)}</span>
+                      <button
+                        onClick={handleDisconnect}
+                        className="bg-transparent border-none text-white cursor-pointer text-sm p-2 transition-colors hover:text-red-400 ml-3"
+                      >
+                        Logout
+                      </button>
+                    </div>
+                    <Button>
+                      <Link to="/dashboard">Go to Dashboard</Link>
+                    </Button>
+                  </div>
+                ) : (
+                  <div className="flex justify-center items-center">
+                    <PlugConnect
+                      whitelist={[process.env.CANISTER_ID_FUTURA_BACKEND]}
+                      onConnectCallback={() => {
+                        console.log("PlugConnect onConnectCallback triggered.");
+                        verifyConnectionAndAgent();
+                      }}
+                    />
+                  </div>
+                )}
+                <p className="px-8 text-center text-sm text-muted-foreground">
+                  By clicking continue, you agree to our{" "}
+                  <span
+                    onClick={() => setView("terms")}
+                    className="underline cursor-pointer underline-offset-4 hover:text-primary"
+                  >
+                    Terms of Service
+                  </span>{" "}
+                  and{" "}
+                  <span
+                    onClick={() => setView("privacy")}
+                    className="underline cursor-pointer underline-offset-4 hover:text-primary"
+                  >
+                    Privacy Policy
+                  </span>
+                  .
+                </p>
+              </div>
+            </div>
           </div>
         </div>
-      </div>
+      )}
+
+      {view === "privacy" && (
+        <div className="p-8">
+          <h2 className="text-2xl font-semibold mb-4">Privacy Policy</h2>
+          <p>{privacyPolicyContent}</p>
+          <button onClick={() => setView("auth")} className="mt-4 underline">
+            Back to Authentication
+          </button>
+        </div>
+      )}
+
+      {view === "terms" && (
+        <div className="p-8">
+          <h2 className="text-2xl font-semibold mb-4">Terms of Service</h2>
+          <p>{termsOfServiceContent}</p>
+          <button onClick={() => setView("auth")} className="mt-4 underline">
+            Back to Authentication
+          </button>
+        </div>
+      )}
     </>
   );
 }
